+++ ticketing-system/migrations/004_create_orders_table.sql
@@ -2,7 +2,7 @@
 CREATE TABLE IF NOT EXISTS orders (
   id UUID PRIMARY KEY DEFAULT gen_random_uuid(),
   user_id UUID NOT NULL REFERENCES users(id) ON DELETE CASCADE,
-  total_amount DECIMAL(10, 2) NOT NULL CHECK (total_amount >= 0),
+  total_amount DOUBLE PRECISION NOT NULL CHECK (total_amount >= 0),
   payment_status VARCHAR(20) NOT NULL CHECK (payment_status IN ('pending', 'completed', 'failed')) DEFAULT 'pending',
   payment_method VARCHAR(50) NOT NULL,
   created_at TIMESTAMP WITH TIME ZONE DEFAULT CURRENT_TIMESTAMP,
